+++ src/configuration/lspConfigurationOptions.ts
@@ -1,5 +1,6 @@
 export const LspConfigurationOptions = [
   'enableRbsSupport',
+  'enableRequiresAncestor',
   'enableWatchman',
   'highlightUntypedCode',
   'highlightUntypedCodeDiagnosticSeverity',