+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 0.2.24
 - New extension icon.
 - Streamline diagnostic messages for improved clarity.
-- Add **Enable Rbs Support** setting.
+- Add **Enable Rbs Support**, **Enable Requires Ancestor** setting.
 
 ## 0.2.23
 - Improve getting started experience: